+++ src/tools/checkpuzzle.hs
@@ -14,8 +14,7 @@ import qualified Parse.PuzzleTypes as T
 import Options.Applicative
 import Control.Monad
 import Data.Maybe
-import qualified Data.Map as Map
-import Data.Monoid
+import qualified Data.Map.Strict as Map
 import Data.List (intercalate, sort)
 
 import System.Exit